Make index_t an uint64_t
While tinkering with the API I realized there was a huge problem:
Returning RIDs from methods like `body_get_shape` is impossible*, as the
library "forgot" the pointer to the boxed index, which is what the
module used.

The most straightforward fix was to encode the indices into a `u64`
instead of boxing them. By encoding them in a deterministic way, it is
possible to always directly use them for hashmap lookups. It also
reduces memory usage, is safer and is just faster in every way.

The only disadvantage is that it limits the amount of elements per array
to 2^32. I do not anticipate this limit to be reached anytime and if it
does, we can always transition to u128 (even though the ABI for that
isn't well-defined).

* It is technically possible, but it would be cumbersome and slow