Add body collision exceptions, free bodies and joints, fix joint orientation
The joint orientation fix needs more investigation, it is very bizarre
that the basis needs to be transposed before assignment.
Note that the "wrong"(?) orientation comes from Godot: the Z axis points
in the wrong direction when checked inside Godot with `printf`
statements.

The motor velocity also has to be inverted, which only adds to the
weirdness.

`cargo fmt` has been run too, which is why some unrelated bits have been
touched.