Remove SpaceHandle, disable rescaling of active colliders
It was adding an unneccesary indirection and required writing quite a
bit of boilerplate in the form of `read`/`modify` and `expect`.

Rescaling of colliders attached to a body has been disabled for now due
to a panic (https://github.com/dimforge/rapier/issues/163). It will be
reenabled later (i.e. it will be rewritten soon anyways).