Improve API generation
The most notable change is the generation of the FFI structs. This will
ensure consistency if any fields are added/removed...

The `method` table has been changed to an array, as JSON tables do not
have a guaranteed order.

`soft_body_set_mesh` has been removed from the API for now as the
`const REF *` was making things difficult and it's currently not
possible to test it anyways (Rapier lacks soft/deformable bodies).

The scripts should be more readable now, though they are still ugly.